+++ zipper/metadata/metadata.go
@@ -13,11 +13,11 @@ type md struct {
 	sync.RWMutex
 	SupportedProtocols       map[string]struct{}
 	ProtocolInits            map[string]func(*zap.Logger, types.BackendV2) (types.BackendServer, *errors.Errors)
-	ProtocolInitsWithLimiter map[string]func(*zap.Logger, types.BackendV2, *limiter.ServerLimiter) (types.BackendServer, *errors.Errors)
+	ProtocolInitsWithLimiter map[string]func(*zap.Logger, types.BackendV2, limiter.ServerLimiter) (types.BackendServer, *errors.Errors)
 }
 
 var Metadata = md{
 	SupportedProtocols:       make(map[string]struct{}),
 	ProtocolInits:            make(map[string]func(*zap.Logger, types.BackendV2) (types.BackendServer, *errors.Errors)),
-	ProtocolInitsWithLimiter: make(map[string]func(*zap.Logger, types.BackendV2, *limiter.ServerLimiter) (types.BackendServer, *errors.Errors)),
+	ProtocolInitsWithLimiter: make(map[string]func(*zap.Logger, types.BackendV2, limiter.ServerLimiter) (types.BackendServer, *errors.Errors)),
 }